% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_explainer.R
\name{deploy_explainer}
\alias{deploy_explainer}
\title{Deploy An Explainer}
\usage{
deploy_explainer(
  exp_name,
  model_package,
  droplet = NA,
  port = 8088,
  deploy = TRUE,
  title = "xai2cloud"
)
}
\arguments{
\item{exp_name}{DALEX explainer object or an .rda filename in your working directory containing the explainer}

\item{model_package}{Name of package used to build the model, eg: "randomForest", "gbm".}

\item{droplet}{If you want to deploy the API locally leave the value as \code{NA}. If you want to deploy to DigitalOcean, use the droplet's ID which can be checked by using \code{analogsea::droplets()}}

\item{port}{Port on which you want your API to be deployed}

\item{deploy}{Boolean telling whether the plumber file is run on set port}

\item{title}{Title to be seen in Swagger}
}
\description{
This function deploys an explainer as an REST API with an corresponding swagger.
A new folder will be created in your working directory containing a plumber R file and an .rda file of the explainer.
Deployment can be done either locally or directly to the DigitalOcean's droplet.
Full guide on setting up DigitalOcean's droplet and deploying to the cloud can be found on package's GitHub page.
}
\examples{
# Create a model
library("ranger")
library("DALEX")
model <- ranger(survived~., data = titanic_imputed)

# Create DALEX explainer
explainer <- explain(model,
                     data = titanic_imputed[,-8],
                     y = titanic_imputed$survived)

# Deploy the API
# If you want the API to deploy automatically, set deploy parameter to TRUE

# Locally
deploy_explainer(explainer, model_package = "ranger",
  port = 8070, deploy=FALSE, title = "Local Example")

# To the cloud
\dontrun{
my_droplets <- analogsea::droplets()

# Choose the correct droplets name - xai2cloudExamples in this case
specific_droplet <- my_droplets$xai2cloudExamples
droplet_id <- specific_droplet$id

deploy_explainer(explainer, model_package = "ranger",
                 droplet = droplet_id, port = 8070,
                 title = "Titanic Example")
}
}
