% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_template_data.R
\name{get_template_data}
\alias{get_template_data}
\title{Get Data From Template}
\usage{
get_template_data(exp_name, title)
}
\arguments{
\item{exp_name}{Explainers object or an .rda file name of an explainer}

\item{title}{Title to be seen in Swagger}
}
\value{
A list of parameters to fill the template
}
\description{
Get Data From Template
}
