% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_load.R
\name{object_load}
\alias{object_load}
\title{Load An File As Object}
\usage{
object_load(file)
}
\arguments{
\item{file}{File name of an .rda file to be loaded as an object.}
}
\description{
Load An File As Object
}
