% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumb_exp.R
\name{plumb_exp}
\alias{plumb_exp}
\title{Plumb Explainer}
\usage{
plumb_exp(exp_name, title = "xai2cloud", droplet = NA, port = 8088)
}
\arguments{
\item{droplet}{}
}
\value{

}
\description{
Plumb Explainer
}
