% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_explainer.R
\name{deploy_explainer}
\alias{deploy_explainer}
\title{Deploy An Explainer}
\usage{
deploy_explainer(exp_name, title = "xai2cloud", droplet = NA, port = 8088)
}
\arguments{
\item{exp_name}{DALEX explainer object or an .rda filename containing the explainer}

\item{title}{Title to be seen in Swagger}

\item{droplet}{If you want to deploy the API locally leave the value as \code{NA}. If you want to deploy to DigitalOcean, use the droplet's ID which can be checked by using \code{analogsea::droplets()}}

\item{port}{Port on which you want your API to be deployed}
}
\description{
This function deploys an explainer as an REST API with an corresponding swagger.
Deployment can be done either locally or directly to the DigitalOcean's droplet.
A new folder will be created in your working directory containing a plumber R file and an .rda file of the explainer.
}
\examples{
\dontrun{
# Using an explainer object, locally
# Load data
titanic <- na.omit(DALEX::titanic)
model_titanic_rf <- randomForest::randomForest(survived == "yes" ~ gender + age + class + embarked +
fare + sibsp + parch,  data = titanic)
explain_titanic_rf <- DALEX::explain(model_titanic_rf,
data = titanic[,-9],
y = titanic$survived == "yes",
label = "Random Forest v7",
colorize = FALSE)

deploy_explainer(explain_titanic_rf, title = "Titanic", port = 8070)

# Using an explainer object, to the cloud
deploy_explainer(explain_titanic_rf, title = "Titanic", droplet = 185232162, port = 8070)

# Using an .rda explainer file, locally
deploy_explainer("explain_titanic_rf.rda", title = "Titanic", port = 8070)

# Using an .rda explainer file, to the cloud
deploy_explainer("explain_titanic_rf.rda", title = "Titanic", droplet = 185232162, port = 8080)
}
}
