% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_explainer.R
\name{deploy_explainer}
\alias{deploy_explainer}
\title{Deploy An Explainer}
\usage{
deploy_explainer(
  exp_name,
  model_package,
  droplet = NA,
  port = 8088,
  deploy = TRUE,
  title = "xai2cloud"
)
}
\arguments{
\item{exp_name}{DALEX explainer object or an .rda filename in your working directory containing the explainer}

\item{model_package}{Name of package used to build the model, eg: "randomForest", "gbm".}

\item{droplet}{If you want to deploy the API locally leave the value as \code{NA}. If you want to deploy to DigitalOcean, use the droplet's ID which can be checked by using \code{analogsea::droplets()}}

\item{port}{Port on which you want your API to be deployed}

\item{deploy}{Boolean telling whether the plumber file is run on set port}

\item{title}{Title to be seen in Swagger}
}
\description{
This function deploys an explainer as an REST API with an corresponding swagger.
A new folder will be created in your working directory containing a plumber R file and an .rda file of the explainer.
Deployment can be done either locally or directly to the DigitalOcean's droplet.
Full guide on setting up DigitalOcean's droplet and deploying to the cloud can be found on package's GitHub page.
}
\examples{
# Using an explainer object, locally
# Load data and create model
model_data <- DALEX::titanic_imputed
titanic_glm <- glm(survived ~ gender + age + fare,
                       data = model_data, family = "binomial")
# Create DALEX explainer
explain_titanic_glm <- DALEX::explain(titanic_glm,
                           data = model_data,
                           y = model_data$survived,
                           label = "glm")
# Deploy the API
# If you want the API to deploy automatically, set deploy parameter to TRUE

# Locally
deploy_explainer(explain_titanic_glm, model_package = "stats",
  title = "Titanic", port = 8070, deploy=FALSE)

# To the cloud
\dontrun{
analogsea::droplets()
deploy_explainer(explain_titanic_glm, model_package = "stats",
  title = "Titanic", droplet = 136232162,
  port = 8080)
}
}
