% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_template_data.R
\name{get_template_data}
\alias{get_template_data}
\title{Get Data From Template}
\usage{
get_template_data(exp_name, model_package, title)
}
\arguments{
\item{exp_name}{DALEX explainer object or an .rda filename in your working directory containing the explainer}

\item{model_package}{Package used to build the model, eg: "randomForest", "gbm"}

\item{title}{Title to be seen in Swagger}
}
\value{
A list of parameters to fill the template
}
\description{
Get Data From Template
}
